Refactored how video player checks if it should show playlist. Added a button in playlist which shows all lists and able to switch which list is shown. Added param for useSyncedLocalState to not update the localstorage if false. Fixed/changed the button which opens the lists popup to toggle when clicking the "main" button.
Refactored how video player checks if it should show playlist. Added a button in playlist which shows all lists and able to switch which list is shown. Added param for useSyncedLocalState to not update the localstorage if false. Fixed/changed the button which opens the lists popup to toggle when clicking the "main" button.