New feature: Create custom feed sections to filter out specific Twitch live streams based on key words.  Slider added to be able to change the order of feeds by dragging an icon, and removed old up/down arrows used to change order. Top/right/bottom/left added as style options for ToolTip. Attr "relative" added to <MyModal/> which mounts the content on the same node as the modal trigger button instead of body (to allow better position relative to the trigger button.) and use the hook useClicksOutside() to close the modal instead of a backdrop when relative is used. Options obj added to UseInput() hook to save the input as number.
New feature: Create custom feed sections to filter out specific Twitch live streams based on key words.

Slider added to be able to change the order of feeds by dragging an icon, and removed old up/down arrows used to change order.
Top/right/bottom/left added as style options for ToolTip.
Attr "relative" added to <MyModal/> which mounts the content on the same node as the modal trigger button instead of body (to allow better position relative to the trigger button.) and use the hook useClicksOutside() to close the modal instead of a backdrop when relative is used.
Options obj added to UseInput() hook to save the input as number.